+++ src/main/java/com/app/msc_project_administrator/user/UserRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
   Optional<User> findByEmail(String email);
 
-  Optional<User> findById(Integer id);
+  Optional<User> findById(Long id);
 //  List<User> findAllById(List<Integer> ids);
 
   List<User> findAllByUserIdIn(List<Integer> ids);
